requirements.txt
pypdf2
python-dotenv
mistralai
python-docx
streamlit-modal
pandas
streamlit
